feat(exporter): fault tolerant exporter
Refactors the exporter to make it fault tolerant. Records are now
batched in memory, and streamed via a transaction producer. Within the
export entry point, the only possible failure is if the batch is full
and cannot be flushed, at which point the current record will be retried
(with back off) ad infinitum.

Flushing the batch makes use of Kafka transactions. This greatly
simplifies error handling as we only need to handle errors at 4 points:

- when initializing transactions for the producer
- when beginning a transaction
- when sending a record
- when committing the transaction

In each case, every error is considered a non recoverable error, with
the exception of: TimeoutException and InterruptException. In the case
of recoverable errors, the operation that failed will be retried as is
with the same producer. In the case of a non recoverable error, it will
be retried with a new producer, and the previous one will be closed. In
either case, the batch is only cleared when it is successfully flushed
(i.e. the transaction is committed successfully), or when the exporter
closes.